% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting.r
\name{createLCMMsetting}
\alias{createLCMMsetting}
\title{setting for LCMM}
\usage{
createLCMMsetting(
  targetCohortId,
  measurementConceptId,
  trajectoryDuration,
  trajectoryBy = "FU",
  trajectoryByUnit = "day",
  valueSelectMethod = "minimum",
  ageAdjusted = FALSE,
  genderAdjusted = FALSE,
  degreeNum = 2:7,
  classNum = 2:7,
  minimumMeasurementCount,
  randomSeed = NULL
)
}
\arguments{
\item{targetCohortId}{Cohort Definition Id}

\item{measurementConceptId}{measurement concept id interested in}

\item{trajectoryDuration}{vector; from time window start to window end; 0:365}

\item{trajectoryBy}{character; "age" or "FU";trajectory will created by; FU : follow-up-time; x axis}

\item{trajectoryByUnit}{character; "day","month","year"; x axis unit}

\item{valueSelectMethod}{character; "minimum" or "maximum"; value select method if multiple measurement values exist}

\item{ageAdjusted}{logical; when you do lcmm, you want to adjust age?}

\item{genderAdjusted}{logical; When you do lcmm, you want to adjust gender?}

\item{degreeNum}{vector; degree of formula}

\item{classNum}{vector; cluster number}

\item{minimumMeasurementCount}{numeric; minimum number of measurement count that each person must have}

\item{randomSeed}{seed number}
}
\description{
setting for LCMM
}
