% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTemporalLongtermData.r
\name{demographicData}
\alias{demographicData}
\alias{demographic}
\title{get demographic data}
\usage{
demographic(useDemographicsAge, useDemographicsGender, connectionDetails,
  cdmDatabaseSchema, resultDatabaseSchema, cohortTable, targetCohortId)
}
\arguments{
\item{useDemographicsAge}{logical, If you want to adjust for your age, this has to be TRUE}

\item{useDemographicsGender}{logical, If you want to adjust for your gender, this has to be TRUE}
}
\description{
get demographic data
}
