% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTemporalLongtermData.r
\name{getTemporalPlpData}
\alias{getTemporalPlpData}
\title{get temporalPlpData for lcmm}
\usage{
getTemporalPlpData(startDay, endDay, daysInterval, connectionDetails,
  cdmDatabaseSchema, resultDatabaseSchema, cohortTable, targetCohortId,
  outcomeCohortId, washoutPeriod = 0, firstExposureOnly = FALSE,
  removeSubjectsWithPriorOutcome = TRUE, priorOutcomeLookback = 99999,
  requireTimeAtRisk = FALSE, minTimeAtRisk = 0,
  riskWindowStart = NULL, addExposureDaysToStart = FALSE,
  riskWindowEnd = NULL, addExposureDaysToEnd = FALSE)
}
\arguments{
\item{startDay}{should be integer}

\item{endDay}{should be integer}

\item{daysInterval}{should be integer}

\item{washoutPeriod}{integer (default = 0)}

\item{firstExposureOnly}{logical (default = FALSE)}

\item{removeSubjectsWithPriorOutcome}{logical (default = TRUE)}

\item{priorOutcomeLookback}{integer (default = 99999)}

\item{requireTimeAtRisk}{logical (default = FALSE)}

\item{minTimeAtRisk}{integer (default = 0)}

\item{riskWindowStart}{integer, if NULL, riskWindowStart = startDay}

\item{addExposureDaysToStart}{logical (default = FALSE)}

\item{riskWindowEnd}{integer, if NULL, riskWindowEnd = endDay}

\item{addExposureDaysToEnd}{logical (default = FALSE)}
}
\description{
get temporalPlpData for lcmm
}
