% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrajectoryData.r
\name{getTrajectoryData}
\alias{getTrajectoryData}
\title{get Longitudinal data for trajectory clustering}
\usage{
getTrajectoryData(
  connectionDetails,
  CDMschema,
  oracleTempSchema = NULL,
  Resultschema,
  cohortTable,
  temporalSetting,
  demographicSetting,
  TargetMeasurementConceptId,
  trajectoryByUnit,
  valueSelectMethod,
  minimumMeasurementCount,
  TargetCohortId
)
}
\arguments{
\item{oracleTempSchema}{}
}
\description{
get Longitudinal data for trajectory clustering
}
