% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTemporalLongtermData.r
\name{longtermPopulation}
\alias{longtermPopulation}
\title{get targetLongtermPopulation data}
\usage{
longtermPopulation(temporalPlpData, targetConceptIds,
  minMeasurementCount = 1, useDemographicsAge = FALSE,
  useDemographicsGender = FALSE, connectionDetails = NULL,
  cdmDatabaseSchema = NULL, resultDatabaseSchema = NULL,
  cohortTable = NULL, targetCohortId = NULL)
}
\arguments{
\item{temporalPlpData}{result of getTemporalPlpData (list)}

\item{targetConceptIds}{target long-term followed conceptIds}

\item{minMeasurementCount}{integer, minimum count of measurement, defualt = 1}

\item{useDemographicsAge}{logical, If you want to adjust for your age, this has to be TRUE}

\item{useDemographicsGender}{logical, If you want to adjust for your gender, this has to be TRUE}

\item{connectionDetails}{If useDemographicsAge or useDemographicsGender is TRUE, then you should fill}

\item{cdmDatabaseSchema}{If useDemographicsAge or useDemographicsGender is TRUE, then you should fill}

\item{resultDatabaseSchema}{If useDemographicsAge or useDemographicsGender is TRUE, then you should fill}

\item{cohortTable}{If useDemographicsAge or useDemographicsGender is TRUE, then you should fill}

\item{targetCohortId}{If useDemographicsAge or useDemographicsGender is TRUE, then you should fill}
}
\description{
get targetLongtermPopulation data
}
