% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmm.r
\name{runLcmm}
\alias{runLcmm}
\title{run lcmm}
\usage{
runLcmm(modelSettings, data)
}
\arguments{
\item{modelSettings}{resultant list from the setLcmm function}

\item{data}{should be vector of integer}

\item{randomArg}{should be vector of integer}

\item{classNum}{should be vector of integer. Default is 2 to 7 (2:7)}

\item{randomSeed}{should be one integer}
}
\description{
run lcmm
}
