% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmm.r
\name{setLcmm}
\alias{setLcmm}
\title{set arguments for lcmm}
\usage{
setLcmm(
  fixedArg = c(2),
  mixtureArg = c(1, 2),
  randomArg = c(1, 2),
  classNum = 2:7,
  randomSeed = 1234
)
}
\arguments{
\item{fixedArg}{should be one integer}

\item{mixtureArg}{should be vector of integer}

\item{randomArg}{should be vector of integer}

\item{classNum}{should be vector of integer. Default is 2 to 7 (2:7)}

\item{randomSeed}{should be one integer}
}
\description{
set arguments for lcmm
}
